refactor(component-cli): separation of concerns
