+++ scripts/component/creators/index.js
@@ -1,0 +1,11 @@
+const newComponent = require('./component');
+const newDoc = require('./doc');
+const newTest = require('./test');
+const editIndexes = require('./editIndexes');
+
+module.exports = {
+  newComponent,
+  newDoc,
+  newTest,
+  editIndexes,
+};
