+++ scripts/component/index.js
@@ -1,9 +1,5 @@
-const fs = require('fs');
-
-const { getPaths } = require('./factory');
-const newComponent = require('./component');
-const newTest = require('./test');
-const newDoc = require('./doc');
+const { getPaths } = require('./utils');
+const { newComponent, newTest, newDoc, editIndexes } = require('./creators');
 
 const component = () => {
   const [name] = process.argv.splice(2);
@@ -12,29 +8,7 @@ const component = () => {
   newComponent(name, paths);
   newTest(name, paths);
   newDoc(name, paths);
-
-  const indexes = [
-    {
-      path: `${paths.base}/index.js`,
-      content: fs.readFileSync(`${paths.base}/index.js`, 'utf-8'),
-    },
-    {
-      path: `${paths.base}/index.native.js`,
-      content: fs.readFileSync(`${paths.base}/index.native.js`, 'utf-8'),
-    },
-    {
-      path: `${paths.doc.native}/index.js`,
-      content: fs.readFileSync(`${paths.doc.native}/index.js`, 'utf-8'),
-    },
-  ];
-
-  indexes.forEach(({ path, content }) => {
-    const newContent = content
-      .replace(/(import.*;\n$)/m, `$1import ${name} from './${name}';\n`)
-      .replace(/};$/m, `  ${name},\n };`);
-
-    fs.writeFileSync(path, newContent, () => {});
-  });
+  editIndexes(name, paths);
 };
 
 component();
